Done borrow-1 within budget, but ...
I think that either the hint could be improved since 'borrow' is not the
key point here, or my solution of using derive Copy+Clone is not the
expected solution, even if within budget.

Either way, the key point is that y has a borrow to x, so when it is
dereferenced to be assigned to z, the struct needs to have Deref
implemented, which it does not in the initial implmentation.

By deriving Clone+Copy, Deref is automatically working since copy types
are not subject to the aliasing rules... at least that's what I think
now.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>