StackVec: Remove confusion between StackVec<'a. T> and StackVec<'a, &'a mut T>
It seems that indeed, in 32ce43c471c891743fd1a87ef511b33a6e379113, when
trying to implement IntoIterator I didn't knew what I was doing and
poisoned the code with StackVec<'a, &'a mut T> instead of Stack<'a, T>.

Probably the lifetimes errors lead me astray.

Remove the aberrant type in favor of the correct one.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>